+++ NEWS.md
@@ -5,6 +5,8 @@
 
 * Added function ddim() to show the data dimensions of a table.
 
+* Added function git.name() to read the name of a Git repository.
+
 * Added function pdeps() to analyze the dependencies of a CRAN package.
 
 * Added function taf.example() to copy an example analysis from the TAF package.